Wire up send funds to provider form
This makes the call to the provider, but the balances in the UI don't update
yet. The templated calls to get balances only happen once, and we need
a place to update the balances in the Svelte app. Most likely on a store
because we show the balances in the header and transactions page.