Add transactions to table and in-progress indicator
We want to update the transaction table when a transaction completes. These
changes are added them to the data structure, but the table doesn't update yet.
Re-worked transaction status a bit to show an in-progress indicator. This won't
be visible until we are making real transactions that take some time.