Add transactions table
We want to display transactions to the user when they have completed. Each
displayed transaction shows localized datetime, destination address, amount,
and messageId.
The destination address and messageId are links to calibration.filscan.io
where users can verify activity on the blockchain. Both are also ellipsed for
long values and on mobile.
Note that the ids in for the table rows use blockheight, but have an artificial
number added temporarily while we use mock data.