Add cosign permission expiration check
Move the cosign permission request out of a store into the transactions
component. We check that we have ucan on initializing the component and
pass in a callback to set it false when it expires.
Also added error handling if the request fails, though it redirects to
the auth lobby and we are unlikely to see an error here.